Gchehade/optimize distribution

+++ objects/objects.go
@@ -31,7 +31,7 @@ type Object struct {
 	CustomMetadata  []CustomMetadata `msgpack:"custom_metadata,omitempty" json:"custom_metadata"`
 	Tags            []string         `msgpack:"tags,omitempty" json:"tags"`
 	Entropy         float64          `msgpack:"entropy,omitempty" json:"entropy"`
-	Distribution    [256]float64     `msgpack:"distribution,omitempty" json:"distribution"`
+	Distribution    [256]byte        `msgpack:"distribution,omitempty" json:"distribution"`
 	Flags           uint32           `msgpack:"flags" json:"flags"`
 }
 
@@ -71,9 +71,9 @@ func (o *Object) AddClassification(analyzer string, classes []string) {
 }
 
 type Chunk struct {
-	Checksum     Checksum     `msgpack:"checksum" json:"checksum"`
-	Length       uint32       `msgpack:"length" json:"length"`
-	Entropy      float64      `msgpack:"entropy" json:"entropy"`
-	Flags        uint32       `msgpack:"flags" json:"flags"`
-	Distribution [256]float64 `msgpack:"distribution,omitempty" json:"distribution"`
+	Checksum     Checksum  `msgpack:"checksum" json:"checksum"`
+	Length       uint32    `msgpack:"length" json:"length"`
+	Entropy      float64   `msgpack:"entropy" json:"entropy"`
+	Flags        uint32    `msgpack:"flags" json:"flags"`
+	Distribution [256]byte `msgpack:"distribution,omitempty" json:"distribution"`
 }
